+++ index.php
@@ -105,7 +105,11 @@ endif;
         <?php if (tplTurquoiseHelper::isFrontpage() && $this->countModules('featured')) : ?>
           <jdoc:include type="modules" name="featured" style="featured" />
         <?php endif; ?>
-          <jdoc:include type="component" />
+          <div class="row">
+            <div class="col-md-12">
+              <jdoc:include type="component" />
+            </div>
+          </div>
         </div>
       <?php if ($this->countModules('right')) : ?>
         <div class="col-md-3">